requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
gunicorn==20.1.0
Jinja2==3.0.1
psycopg2-binary==2.9.2
pymongo==3.12.0
python-dateutil==2.8.2
pytz==2021.1
virtualenv==20.8.1
virtualenvwrapper-win==1.2.6
Werkzeug==2.0.1
